% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwiki-package.R
\docType{package}
\name{rwiki-package}
\alias{rwiki}
\alias{rwiki-package}
\title{rwiki: Download wikibook in R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
