% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibook.R
\name{wikibook}
\alias{wikibook}
\title{wikibook}
\usage{
wikibook(url, lang = "en")
}
\arguments{
\item{url}{the url of wiki book}

\item{lang}{language}
}
\description{
Retrieve wiki information
}
\examples{
url  <- "https://en.wikipedia.org/wiki/User:Kongdd/Books/\%E7\%BB\%9F\%E8\%AE\%A1\%E5\%AD\%A6\%E5\%9F\%BA\%E7\%A1\%80"
info <- wikibook(url)

outdir <- info$origin$title[1]
# plyr::m_ply(info$origin[, .(url = url, outfile = title)],
#             write_pdf, outdir = outdir)
# plyr::m_ply(info$lang[, .(url = url, outfile = title)],
#             write_pdf, outdir = outdir)
}
